fix args for create/update file, and fix new body for update_file
The content passed into update_file is not appended - it is an overwrite of the file. So, we have to manually append onto the existing content.

Note that thea the new target branch git ref has to exist for either type of file change to be committed successfully to the cloud